Add Preliminary Login and Register Pages.
These files are being committed very late, and the work in this commit is
unfinished. A server is set up with Passport authenication set up. Certain
paths are public, whereas others require a JWT to view. However, at the
moment, Login functionality does not work because the submit action on the
front-end is unhandled. Theoretically, if that is addded, this application
should serve as a basic login functionality.